// src/app/interfaces/tag.ts
export interface Tag {
    id_tag: number;
    final_tag: string;
    asset_number: string;
    edifice: string;
    floor: string;
    sector: string;
    site: string;
    asset_type: string;
    id_edifice: number;
    id_floor: number;
    id_sector: number;
    id_site: number;
    id_asset_type: number;
}