// src/app/interfaces/task-list-ot.ts
export interface TaskListOt {
    id_task_list: number;
    asset_type_name: string; // Cambiamos a asset_type_name para que coincida con los datos
    task_type_name: string; // Cambiamos a task_type_name para que coincida con los datos
    step_1: number;
    step_2: number;
    step_3: number;
    step_4: number;
    step_5: number;
    step_6: number;
    step_7: number;
    step_8: number;
    step_9: number;
    step_10: number;
  }